check installed

+++ debian/setup_go.sh
@@ -1,10 +1,15 @@
 #!/bin/bash
 
 echo "install go 1.18"
+echo "install go"
+if [ -x "$(command -v go)" ]; then
+  echo "go is already installed, exit."
+  exit 1
+fi
 
 go_bin_filename="go1.18.linux-amd64.tar.gz"
 wget https://golang.org/dl/${go_bin_filename}
-sudo tar -zxvf ${go_bin_filename} -C /usr/local/
+sudo tar -zxvf ${go_bin_filename} -C /usr/local/ >> /dev/null
 rm ${go_bin_filename}
 # todo: a better place to set $PATH
 # bash
